+++ src/VirtoCommerce.PageBuilderModule.Web/page-builder/src/modules/page-builder/components/index.ts
@@ -2,5 +2,7 @@ import PageStatus from "./pageStatus.vue";
 import CardHeader from "./cardHeader.vue";
 import SwitchRow from "./switchRow.vue";
 import Tag from "./tag.vue";
+import PagesList, { ExposedPagesList } from "./pagesList.vue";
 
-export { PageStatus, CardHeader, SwitchRow, Tag };
+export { PageStatus, CardHeader, SwitchRow, Tag, PagesList };
+export type { ExposedPagesList };
