+++ src/VirtoCommerce.PageBuilderModule.Web/page-builder/src/modules/page-builder/composables/index.ts
@@ -1,6 +1,5 @@
-export { usePageBuilderList } from "./usePageBuilderList";
+export { PageStatuses, usePageBuilderList } from "./usePageBuilderList";
 export { usePageBuilderDetails } from "./usePageBuilderDetails";
 export { default as useCultureNames } from "./useCultureNames";
 export { default as useUserGroups } from "./useUserGroups";
 export { default as useUrlParams } from "./useUrlParams";
-
