+++ src/VirtoCommerce.PageBuilderModule.Web/page-builder/src/modules/page-builder/composables/usePagesListToolbar/index.ts
@@ -1,0 +1,41 @@
+import { computed, Ref, ShallowRef } from "vue";
+import { useI18n } from "vue-i18n";
+import { IBladeToolbar } from "@vc-shell/framework";
+import { PageStatuses } from "../usePageBuilderList";
+import { ExposedPagesList } from "../../components";
+
+export function usePagesListToolbar(
+  status: PageStatuses | null,
+  pagesListRef: Readonly<ShallowRef<ExposedPagesList | null>>,
+): Ref<IBladeToolbar[]> {
+
+  const { t } = useI18n({ useScope: "global" });
+
+  return computed(() => [
+    {
+      id: "add",
+      icon: "material-add",
+      title: t("PAGE_BUILDER.PAGES.LIST.TOOLBAR.ADD"),
+      clickHandler: async () => {
+        await pagesListRef.value?.openAddBlade();
+      },
+    },
+    {
+      id: "refresh",
+      icon: "material-refresh",
+      title: t("PAGE_BUILDER.PAGES.LIST.TOOLBAR.REFRESH"),
+      clickHandler: async () => {
+        await pagesListRef.value?.reload();
+      },
+    },
+    {
+      id: "delete",
+      icon: "material-delete",
+      title: t("PAGE_BUILDER.PAGES.LIST.TOOLBAR.REMOVE"),
+      disabled: pagesListRef.value?.selectedItems.value?.length === 0,
+      clickHandler: async () => {
+        await pagesListRef.value?.removeSelectedPages();
+      },
+    },
+  ]);
+}
