+++ src/VirtoCommerce.PageBuilderModule.Web/page-builder/src/modules/page-builder/pages/index.ts
@@ -1,3 +1,3 @@
-export { default as AllPages } from "./AllPages.vue";
-export { default as ArchivedPages } from "./ArchivedPages.vue";
+export { default as AllPagesList } from "./AllPagesList.vue";
+export { default as ArchivedPagesList } from "./ArchivedPagesList.vue";
 export { default as PageDetails } from "./PageDetails.vue";
